refactor: introduce shared utils and optimize battle
